+++ src/main/java/com/iti/jets/reportingsystem/services/impls/IntervalsInfoServiceImpl.java
@@ -4,6 +4,7 @@ import com.iti.jets.openapi.model.IntervalsInfoRequest;
 import com.iti.jets.openapi.model.IntervalsInfoResponse;
 import com.iti.jets.reportingsystem.entities.IntervalsInfo;
 import com.iti.jets.reportingsystem.entities.Well;
+import com.iti.jets.reportingsystem.exceptions.ResourceNotFoundException;
 import com.iti.jets.reportingsystem.repos.IntervalsInfoRepository;
 import com.iti.jets.reportingsystem.repos.WellRespository;
 import com.iti.jets.reportingsystem.services.IntervalsInfoService;
@@ -55,6 +56,10 @@ public class IntervalsInfoServiceImpl implements IntervalsInfoService {
     }
 
     public IntervalsInfoResponse getIntervalsInfoById(int id) {
+        if(!intervalsInfoRepository.findById(id).isPresent())
+        {
+            throw new ResourceNotFoundException("There is no record with this id");
+        }
         IntervalsInfo intervalsInfo = intervalsInfoRepository.findById(id).get();
         if (intervalsInfo != null) {
             IntervalsInfoResponse intervalsInfoResponse = mapper.map(intervalsInfoRepository.findById(id).get(), IntervalsInfoResponse.class);
@@ -92,6 +97,10 @@ public class IntervalsInfoServiceImpl implements IntervalsInfoService {
     }
 
     public boolean updateIntervalsInfo(int id, IntervalsInfoRequest intervalsInfoRequest) {
+        if(!intervalsInfoRepository.findById(id).isPresent())
+        {
+            throw new ResourceNotFoundException("There is no record with this id");
+        }
         if (intervalsInfoRepository.findById(id).isPresent()) {
             IntervalsInfo intervalsInfo = mapper.map(intervalsInfoRequest, IntervalsInfo.class);
 
@@ -112,7 +121,10 @@ public class IntervalsInfoServiceImpl implements IntervalsInfoService {
     }
 
     public boolean deleteIntervalsInfo(int id) {
-
+         if(!intervalsInfoRepository.findById(id).isPresent())
+         {
+             throw new ResourceNotFoundException("There is no record with this id");
+         }
         if (intervalsInfoRepository.findById(id).isPresent()) {
             intervalsInfoRepository.deleteById(id);
             return true;
